+++ drawingnote_application/lib/bluetooth_connecting_page.dart
@@ -1,0 +1,170 @@
+import 'package:bluetooth_classic/models/device.dart';
+import 'package:flutter/material.dart';
+import 'dart:async';
+import 'dart:convert';
+
+import 'package:flutter/services.dart';
+import 'package:bluetooth_classic/bluetooth_classic.dart';
+
+class BluetoothConnectingPage extends StatefulWidget {
+  const BluetoothConnectingPage({super.key});
+
+  @override
+  State<BluetoothConnectingPage> createState() =>
+      _BluetoothConnectingPageState();
+}
+
+class _BluetoothConnectingPageState extends State<BluetoothConnectingPage> {
+  //bluetooth 관련 변수 선언
+  String _platformVersion = 'Unknown';
+  final _bluetoothClassicPlugin = BluetoothClassic();
+  List<Device> _devices = [];
+  List<Device> _discoveredDevices = [];
+  bool _scanning = false;
+  int _deviceStatus = Device.disconnected;
+
+  //final Uint8List _data = Uint8List(0);
+  final List<String> _receivedInput = [];
+
+  //bluetooth 관련 함수 선언 (DeviceStatus 변경, 데이터 receive)
+  @override
+  void initState() {
+    super.initState();
+    initPlatformState();
+    _bluetoothClassicPlugin.onDeviceStatusChanged().listen((event) {
+      setState(() {
+        _deviceStatus = event;
+      });
+    });
+    _bluetoothClassicPlugin.onDeviceDataReceived().listen((event) {
+      setState(() {
+        _receivedInput.add(utf8.decode(event));
+        //_data = Uint8List.fromList([..._data, ...event]);
+      });
+    });
+  }
+
+  //기기 OS version 받아오기
+  // Platform messages are asynchronous, so we initialize in an async method.
+  Future<void> initPlatformState() async {
+    String platformVersion;
+    // Platform messages may fail, so we use a try/catch PlatformException.
+    // We also handle the message potentially returning null.
+    try {
+      platformVersion = await _bluetoothClassicPlugin.getPlatformVersion() ??
+          'Unknown platform version';
+    } on PlatformException {
+      platformVersion = 'Failed to get platform version.';
+    }
+
+    // If the widget was removed from the tree while the asynchronous platform
+    // message was in flight, we want to discard the reply rather than calling
+    // setState to update our non-existent appearance.
+    if (!mounted) return;
+
+    setState(() {
+      _platformVersion = platformVersion;
+    });
+  }
+
+  //paired device 가져와서 _devices에 저장
+  Future<void> _getDevices() async {
+    var res = await _bluetoothClassicPlugin.getPairedDevices();
+    setState(() {
+      _devices = res;
+    });
+  }
+
+  //bluetooth device 스캔 해서 _discoveredDevices에 저장
+  Future<void> _scan() async {
+    if (_scanning) {
+      await _bluetoothClassicPlugin.stopScan();
+      setState(() {
+        _scanning = false;
+      });
+    } else {
+      await _bluetoothClassicPlugin.startScan();
+      _bluetoothClassicPlugin.onDeviceDiscovered().listen(
+        (event) {
+          setState(() {
+            _discoveredDevices = [..._discoveredDevices, event];
+          });
+        },
+      );
+      setState(() {
+        _scanning = true;
+      });
+    }
+  }
+
+  //화면 구성
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      home: Scaffold(
+        appBar: AppBar(
+          title: const Text('Plugin example app'),
+        ),
+        body: SingleChildScrollView(
+          child: Column(
+            children: [
+              Text("Device status is $_deviceStatus"),
+              TextButton(
+                onPressed: () async {
+                  await _bluetoothClassicPlugin.initPermissions();
+                },
+                child: const Text("Check Permissions"),
+              ),
+              TextButton(
+                onPressed: _getDevices,
+                child: const Text("Get Paired Devices"),
+              ),
+              TextButton(
+                onPressed: _deviceStatus == Device.connected
+                    ? () async {
+                        await _bluetoothClassicPlugin.disconnect();
+                      }
+                    : null,
+                child: const Text("disconnect"),
+              ),
+              TextButton(
+                onPressed: _deviceStatus == Device.connected
+                    ? () async {
+                        await _bluetoothClassicPlugin.write("ping\r\n");
+                      }
+                    : null,
+                child: const Text("send ping"),
+              ),
+              Center(
+                child: Text('Running on: $_platformVersion\n'),
+              ),
+              ...[
+                for (var device in _devices)
+                  TextButton(
+                      onPressed: () async {
+                        await _bluetoothClassicPlugin.connect(device.address,
+                            "00001101-0000-1000-8000-00805f9b34fb");
+                        setState(() {
+                          _discoveredDevices = [];
+                          _devices = [];
+                        });
+                      },
+                      child: Text(device.name ?? device.address))
+              ],
+              TextButton(
+                onPressed: _scan,
+                child: Text(_scanning ? "Stop Scan" : "Start Scan"),
+              ),
+              ...[
+                for (var device in _discoveredDevices)
+                  Text(device.name ?? device.address)
+              ],
+              //Text("Received data: ${String.fromCharCodes(_data)}"),
+              for (String input in _receivedInput) Text(input),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
